Android: make sure that a URL gets open in our app rather than in user's preferred browser.
Although I believe this is the correct way to do it, it doesn't seem to be working properly. We are getting a white screen to start with and, then, if we wait long enough, we eventually get a message that says that the webpage is not available and we get an ERR_NAME_NOT_RESOLVED error. I imagine it must be an issue with the Android VM?...