Added support for notifications.
I believe I have done everything there is to have support for notifications and it's all working fine on iOS (both emulator and iPhone), but not on our Android emulator (don't have an Android phone). Could it be due to my Android emulator not working properly (i.e. no access to the Internet through flutter)!?