Log in: only do so through the WaniKani website now.
Indeed, it (obviously) works fine if the user provides his/her API token, but... this means that we can't retrieve his/her email address and therefore can't retrieve his/her gravatar. So, to have two different ways of logging in that result in different gravatars is not good, hence only one way of logging in now.